narrow query manager exports and split runtime core tests
